% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRETIGEA_package.R
\docType{data}
\name{aba_marker_expression}
\alias{aba_marker_expression}
\title{Normalized FPKM expression data from a subset of the Allen Brain Atlas Aging, Dementia, and TBI study.}
\format{
An object of class \code{data.frame} with 395 rows and 377 columns.
}
\usage{
aba_marker_expression
}
\description{
Filtered data set to only contain brain marker genes as well as ~100 other genes. The total data set can be downloaded by following the links in the original paper.
}
\references{
© 2016 Allen Institute for Cell Science. Aging, Dementia and TBI. Available from: http://aging.brain-map.org/
}
\keyword{datasets}
