% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRETIGEA_package.R
\docType{data}
\name{aba_pheno_data}
\alias{aba_pheno_data}
\title{Phenotype data from brain samples in the Allen Brain Atlas Aging, Dementia, and TBI study.}
\format{
An object of class \code{data.frame} with 377 rows and 4 columns.
}
\usage{
aba_pheno_data
}
\description{
A subset of phenotype data to be used for correlations with the expression data in the tests and vignette, to validate that the estimated cell type proportions correspond to useful quantities.
}
\references{
© 2016 Allen Institute for Cell Science. Aging, Dementia and TBI. Available from: http://aging.brain-map.org/
}
\keyword{datasets}
