% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRETIGEA_package.R
\docType{data}
\name{kelley_df_brain}
\alias{kelley_df_brain}
\title{Marker genes estimated from}
\format{
An object of class \code{data.frame} with 84000 rows and 2 columns.
}
\usage{
kelley_df_brain
}
\description{
Top 1000 marker genes from each of four major brain cell types (ie astrocytes, microglia, neurons, and oligodendrocytes) estimated via correlation-based analysis of bulk brain tissue. Contains marker data averaged across all brain regions as well as markers derived from individual brain regions.
}
\references{
Kelley KW, Nakao-inoue H, Molofsky AV, Oldham MC. Variation among intact tissue samples reveals the core transcriptional features of human CNS cell classes. Nat Neurosci. 2018;21(9):1171-1184.
}
\keyword{datasets}
