% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRETIGEA_package.R
\docType{data}
\name{markers_df_human_brain}
\alias{markers_df_human_brain}
\title{Marker genes estimated from a meta-analysis of brain cell gene expression data from humans only.}
\format{An object of class \code{data.frame} with 5500 rows and 2 columns.}
\usage{
markers_df_human_brain
}
\description{
Top 1000 marker genes from each of the six major brain cell types (ie astrocytes, endothelial cells, microglia, neurons, oligodendrocytes, and OPCs) estimated from a meta-analysis of brain cell gene expression data from humans only.
}
\references{
Mckenzie AT, Wang M, Hauberg ME, et al. Brain Cell Type Specific Gene Expression and Co-expression Network Architectures. Sci Rep. 2018;8(1):8868.
}
\keyword{datasets}
