% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRETIGEA_package.R
\docType{package}
\name{BRETIGEA}
\alias{BRETIGEA}
\title{BRETIGEA: BRain cEll Type specIfic Gene Expression Analysis}
\description{
This package provides two major functions:
1) A function to estimate cell type surrogate proportion variables based on marker genes.
2) A function to adjust bulk gene expression data for estimated cell type proportions, as covariates.
It also provides data containing estimated marker genes for six major brain cell types, i.e. astrocytes, endothelial cells, microglia, neurons, oligodendrocytes, and oligodendrocyte precursor cells (OPCs), as well as wrapper functions to use this data on the two major package functions.
}
